% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyclone.R
\name{calculation_barth_muschelknautz}
\alias{calculation_barth_muschelknautz}
\title{Cyclone Simulation: Barth/Muschelknautz}
\usage{
calculation_barth_muschelknautz(cyclone, fluid, xmean, delta)
}
\arguments{
\item{cyclone}{vector of cyclone's geometrical parameters (Da, Dt, H, Ht, He, Be)}

\item{fluid}{list of fluid parameters}

\item{xmean}{vector of middle points for the intervals of the particle size distribution}

\item{delta}{vector, amount of dust (percentage value) in each interval}
}
\value{
a vector (-collection efficiency, pressure drop)
}
\description{
Calculate cyclone collection efficiency according to Barth/Muschelknautz.
}
