% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpoints.R
\docType{package}
\name{Checkpoint1}
\alias{Checkpoint1}
\title{Checkpoint 1}
\description{
The first stage of the Cyclone pipeline
}
\details{
In this stage, the pipeline reads in raw files and prepares the data for umap calculation.
}
\examples{
\dontrun{
In a terminal window, run:
Rscript cytof_pipeline.R my_config.yml
}
}
