% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cmops}
\alias{cmops}
\title{CMOP bounds}
\format{
A list of 3 CMOPs and 6 variables representing cyclone's geometrical parameters (Da, Dt, H, Ht, He, Be)
}
\usage{
cmops
}
\description{
The dataset containing bounds of CMOPs.
}
\keyword{datasets}
