% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_cmop}
\alias{create_cmop}
\title{Creating problem instance}
\usage{
create_cmop(cyclone, eps = 0.1)
}
\arguments{
\item{cyclone}{Vector of default cyclone's geometrical parameters (Da, Dt, H, Ht, He, Be).}

\item{eps}{Float from [0, 1] denoting variance in gemoetrical parameters. Default is 0.1.}
}
\value{
A list of upper and lower bounds.
}
\description{
This function creates problem instances based on default cyclone parameters
}
