% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{cyclone-package}
\alias{cyclone-package}
\title{Cyclone Toolbox}
\description{
Cyclone Toolbox
}
\details{
Toolbox for multiobjective optimization of cyclone dust separator.

\tabular{ll}{
Package: \tab cyclone\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2020-06-17\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\section{Maintainer}{

Aljosa Vodopija \email{aljosa.vodopija@ijs.si}
}

\author{
Aljosa Vodopija \email{aljosa.vodopija@ijs.si}
}
