% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/followUp.R
\name{downsample_sce}
\alias{downsample_sce}
\title{A function to create a down-sampled SingleCellExperiment object from a full one.}
\usage{
downsample_sce(full_sce, n_keep = 1e+05, verbose = TRUE)
}
\arguments{
\item{full_sce}{the \code{\link[SingleCellExperiment]{SingleCellExperiment}} object to downsample}

\item{n_keep}{Positive integer. The number of cells to retain.}

\item{verbose}{Logical. Whether to output timestamped log messages during running.}
}
\value{
a \code{\link[SingleCellExperiment]{SingleCellExperiment}} object
}
\description{
A function to create a down-sampled SingleCellExperiment object from a full one.
}
\details{
\code{full_sce} is subset to a randomly selected \code{n_keep} number of cells, and then returned.
}
\section{NOTE}{

performs a simple downsample that does NOT attempt to pull equally from each
sample or cluster.  The purpose here is assumed to simply be rapid testing of
visualizations which can take minutes longer to produce from millions of
cells than from thousands of cells.
}

\author{
Daniel Bunis
}
