% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{feas_ratios}
\alias{feas_ratios}
\title{Feasibility ratios}
\usage{
feas_ratios(problem, sample.size = 1e+06)
}
\arguments{
\item{problem}{A list representing the problem configuration:
\describe{
    \item{\code{lower.bounds}}{Lower box constraints.}
    \item{\code{upper.bounds}}{Upper box constraints.}
    \item{\code{cons}}{Constraints considered in optimization.}
}}

\item{sample.size}{An integer denoting the sample size.}
}
\value{
A vector of feasibility ratios per constraints and the overall feasibility ratio.
}
\description{
This function calculates feasibility ratios.
}
