% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyclone.R
\name{fun_cyclone}
\alias{fun_cyclone}
\title{Objective function - Cyclone Simulation: Barth/Muschelknautz}
\usage{
fun_cyclone(
  cyclone,
  fluid = NULL,
  intervals = c(0, 2, 4, 6, 8, 10, 15, 20, 30) * 1e-06,
  delta = c(0, 0.02, 0.03, 0.05, 0.1, 0.3, 0.3, 0.2),
  cons.bounds = c(0.9, 1500)
)
}
\arguments{
\item{cyclone}{vector of cyclone's geometrical parameters (Da, Dt, H, Ht, He, Be)}

\item{fluid}{list of default fluid parameters (Mu, Ve, lambdag, Rhop, Rhof, Croh)}

\item{intervals}{vector specifying the particle size interval bounds}

\item{delta}{vector, amount of dust (percentage value) in each interval}

\item{cons.bounds}{vector of objective bounds for (CEmin, PDmax)}
}
\value{
a vector (-CE, PD, geometrical constraints, CE constraint, PD constraint)
}
\description{
Calculate cyclone collection efficiency. A simple, physics-based
optimization problem (potentially bi-objective). See the references [1,2].
}
\references{
[1] Zaefferer, M.; Breiderhoff, B.; Naujoks, B.; Friese, M.; Stork, J.; Fischbach, A.; Flasch, O.; Bartz-Beielstein, T. Tuning Multi-objective Optimization Algorithms for Cyclone Dust Separators Proceedings of the 2014 Conference on Genetic and Evolutionary Computation, ACM, 2014, 1223-1230 \cr\cr
[2] Breiderhoff, B.; Bartz-Beielstein, T.; Naujoks, B.; Zaefferer, M.; Fischbach, A.; Flasch, O.; Friese, M.; Mersmann, O.; Stork, J.; Simulation and Optimization of Cyclone Dust Separators Proceedings 23. Workshop Computational Intelligence, 2013, 177-196
}
