% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/followUp.R
\name{make_or_load_downsample_sce}
\alias{make_or_load_downsample_sce}
\title{A function to create a down-sampled SingleCellExperiment object from a full one OR to load one in that has already been generated and saved.}
\usage{
make_or_load_downsample_sce(
  down_sce_file_name,
  full_sce,
  n_keep = 1e+05,
  save = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{down_sce_file_name}{String. File path of either the sce to load in, or where to save to when \code{save = TRUE}.}

\item{full_sce}{the \code{\link[SingleCellExperiment]{SingleCellExperiment}} object to downsample}

\item{n_keep}{Positive integer. The number of cells to retain.}

\item{save}{Logical. Whether to save the SCE object if newly creating it.}

\item{verbose}{Logical. Whether to output timestamped log messages during running.}
}
\value{
a \code{\link[SingleCellExperiment]{SingleCellExperiment}} object
}
\description{
A function to create a down-sampled SingleCellExperiment object from a full one OR to load one in that has already been generated and saved.
}
\details{
If a file exists at the specified \code{down_sce_file_name} location, that file will be loaded in and returned.
Otherwise, \code{full_sce} will be subset to a randomly selected \code{n_keep} number of cells.
Then, if \code{save} is \code{TRUE}, \code{\link{save_sce}} is then used to save the downsampled SCE to the \code{down_sce_file_name} file path.
Finally, the downsampled SCE is returned.
}
\section{NOTE}{

performs a simple downsample that does NOT attempt to pull equally from each
sample or cluster.  The purpose here is assumed to simply be rapid testing of
visualizations which can take minutes longer to produce from millions of
cells than from thousands of cells.
}

\author{
Daniel Bunis
}
