% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/followUp.R
\name{make_or_load_full_sce}
\alias{make_or_load_full_sce}
\title{A function to automatically import cyclone outputs into a SingleCellExperiment structure OR load one that has already been generated and saved.}
\usage{
make_or_load_full_sce(
  sce_file_name,
  checkpoint1,
  checkpoint8,
  load_checkpoints = TRUE,
  save = FALSE,
  make_clusters_factors = TRUE,
  verbose = TRUE,
  verbose_checkpoint_load = TRUE
)
}
\arguments{
\item{sce_file_name}{String. File path of either the sce to load in, or where to save to when \code{save = TRUE}.}

\item{checkpoint1, checkpoint8}{Strings. File paths pointing to cyclone outputs 'checkpoint_1.RData' and 'checkpoint_8.RData' to use.}

\item{load_checkpoints}{Logical. Whether or not to load in checkpoint1 and checkpoint8 data.}

\item{save}{Logical. Whether to save the SCE object if newly creating it here from checkpoint data}

\item{make_clusters_factors}{Logical. When creating the SCE object, whether to ensure 'cluster' metadata are factors with levels in numeric order.
Doing so ensures clusters appear in order from 1,2,3,4,5, etc. when plotting.}

\item{verbose}{Logical. Whether to output timestamped log messages during running.}

\item{verbose_checkpoint_load}{= Logical. Whether to set 'verbose = TRUE' in \code{\link[base]{load}} calls for loading 'checkpoint1' and 'checkpoint8' data.}
}
\value{
a \code{\link[SingleCellExperiment]{SingleCellExperiment}} object
}
\description{
A function to automatically import cyclone outputs into a SingleCellExperiment structure OR load one that has already been generated and saved.
}
\details{
The function starts by loading in primary outputs of the cyclone pipeline if \code{load_checkpoints = TRUE}.
When doing so, it will read in \code{checkpoint1} before \code{checkpoint8},
a necessary ordering because the cell_metadata element of \code{checkpoint1} is updated to include clustering information by \code{checkpoint8}.
Next, if a file exists at the specified \code{sce_file_name} location, that file will be loaded in and is assumed to be an SCE previously created from the cyclone data.
Otherwise, a new SingleCellExperiment (SCE) object is created where: \itemize{
\item The 'trans_exp' output from checkpoint1 is used to fill a 'transformed' assay.
\item The 'cell_metadata' output from checkpoint8 is used to fill both  colData and a reducedDim named 'umap'.
}
When \code{make_clusters_factors} is \code{TRUE}, all colData columns whose names start with 'column' are turned into factors with levels ordered from min value to max value.
When \code{save} is \code{TRUE} and an SCE was newly created, \code{\link{save_sce}} is then used to save the newly created SCE to the \code{sce_file_name} file path.
Finally, the SCE is returned.
}
\author{
Daniel Bunis
}
