% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{make_plot}
\alias{make_plot}
\title{Making plots}
\usage{
make_plot(res, run = 1)
}
\arguments{
\item{res}{Output from \code{opt_mo}.}

\item{run}{An integer denoting which run to depict in the plot.}
}
\value{
A plot depicting Pareto front approximation of the selected run.
}
\description{
This function makes a plot of the selected Pareto front approximation.
}
