% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{opt_mo}
\alias{opt_mo}
\title{Optimization}
\usage{
opt_mo(problem, method = c("nsga2", "demo"), control = NULL, no.runs = 1)
}
\arguments{
\item{problem}{A list representing the problem configuration:
\describe{
    \item{\code{lower.bounds}}{Lower box constraints.}
    \item{\code{upper.bounds}}{Upper box constraints.}
    \item{\code{cons}}{Constraints considered in optimization.}
}}

\item{method}{A string indicating the used optimization algorithm.
It can be NSGA-II ("nsga2"), DEMO ("demo") or MOEA/D ("moead"). Default is DEMO.}

\item{control}{A list of control parameters:
#' \describe{
    \item{\code{pop.size}}{An integer denoting the number of solutions
    used by the optimization algorithm. Default is 100.}
    \item{\code{no.iters}}{An integer denoting the number of iterations
    (generations) used by the optimization algorithm. Default is 100.}
    \item{\code{cross.prob}}{A float from [0,1] denoting the crossover
    probability. Default is 0.9.}
    \item{\code{mut.prob}}{A float from [0,1] denoting the mutation probability.
    For DEMO this is the scaling factor. Default is 0.1 for NSGA-II and MOEA/D,
    while 0.7 for DEMO.}
}
Default values are used for unspecified parameters.}

\item{no.runs}{An integer denoting the number of runs. Default is 1.}
}
\value{
List of data frames storing the objective values of all
nonodominated feasible solutions found during each run.
}
\description{
This function solves the cyclone optimization problem
}
