% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{param_stat}
\alias{param_stat}
\title{Computing parameter statistics}
\usage{
param_stat(res, run = 1, dec = 8)
}
\arguments{
\item{res}{Output from \code{opt_mo}.}

\item{run}{An integer denoting which run to use.}

\item{dec}{An integer denoting number of decimal points.}
}
\value{
Means and deviations of geometrical parameters.
}
\description{
This function computes the mean and sd of parameters given a set of solutions
}
