% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/followUp.R
\name{save_sce}
\alias{save_sce}
\title{Minor wrapper on top of saveRDS that automatically sets whether to compress the object based on cell number.}
\usage{
save_sce(sce, sce_file_name, compress = NA, verbose, ...)
}
\arguments{
\item{sce}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}} object}

\item{sce_file_name}{String. A filepath at which to save the \code{sce}.}

\item{compress}{Logical or NA. Whether to compress the saved object per like-named input of \code{\link[base]{readRDS}}. When left as \code{NA}, compression is performed if the \code{sce} contains more than 1,000,000 cells.}

\item{verbose}{Logical. Whether to output timestamped log messages during running.}

\item{...}{Additional arguments passed to \code{\link[base]{readRDS}}}
}
\value{
None
}
\description{
Minor wrapper on top of saveRDS that automatically sets whether to compress the object based on cell number.
}
\author{
Daniel Bunis
}
