% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_checkpoint1.R
\name{prepare_checkpoint1}
\alias{prepare_checkpoint1}
\title{prepare_checkpoint1}
\usage{
prepare_checkpoint1(
  raw_exp = NULL,
  trans_exp = NULL,
  file_metadata,
  marker_metadata,
  cell_metadata,
  out_dir,
  arcsinh_cofactor = 5
)
}
\arguments{
\item{raw_exp}{data frame of raw expression (cell x features), optional if trans_exp is provided.}

\item{trans_exp}{data frame of processed (transformed and/or normalized) expression (cell x features), optional if raw_exp is provided. If only raw_exp is provided, the trans_exp is generated by performing arcsinh transformation (default cofactor 5). If trans_exp is provided, the trans_exp is used as it is without any additional transformation/normalization are performed.}

\item{file_metadata}{data frame describing individual files/samples, one per row. Must have following columns:
file_name: unique names of file/sample
donor_id: identifier indicating the donor of the sample (used only for visualization)
pool_id: CyTOF pool id or batch id (used only for visualization)
control_sample: TRUE/FALSE indicating whether the file/sample is a batch control, as often used in CyTOF.  (used only for visualization when starting the pipeline from step 2)
"any additional metadata columns" (not used only for anything, but are just supplied in the output objects for easy access)}

\item{marker_metadata}{data frame listing all markers/features, one per row. Must have following columns:
marker_name: marker names, must match the column names of raw_exp and trans_exp
used_for_UMAP: TRUE/FALSE indicating whether the marker should be used for UMAP dimensionality reduction.
used_for_clustering: TRUE/FALSE indicating whether the marker should be used for clustering.
used_for_scaffold: TRUE/FALSE indicating whether the marker should be used for SCAFFoLD map generation.}

\item{cell_metadata}{data frame with row names matching the rows of raw_exp and trans_exp. Must have following columns:
file_name: file/sample name matching one of the file_name values of file_metadata}

\item{out_dir}{path of output directory that will be used as out_dir for running the cyclone pipeline.}

\item{arcsinh_cofactor}{cofactor to use during ArcSinh transformation. If trans_exp is provided, this value is ignored.}
}
\description{
Checks if the data is provided in a correct format and generates checkpoint_1.RData in out_dir
}
\author{
Ravi K. Patel
}
